requirements.txt
aiofiles
aiohttp
hikari 
fastapi
uvicorn
asyncpg 
jinja2
dominate
python-dotenv